requirements.txt
affine==2.4.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
git-filter-repo==2.45.0
numpy==2.1.1
pillow==10.4.0
pyparsing==3.1.4
pyproj==3.6.1
rasterio==1.3.11
snuggs==1.4.7
